First patch. MRuby & mruby-apr now at least able to run tests.

+++ lib/mspec/helpers/tmp.rb
@@ -9,25 +9,37 @@ SPEC_TEMP_UNIQUIFIER = "0"
 
 SPEC_TEMP_DIR_PID = Process.pid
 
-at_exit do
-  begin
-    if SPEC_TEMP_DIR_PID == Process.pid
-      Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+if Kernel.respond_to?(:at_exit)
+  at_exit do
+    begin
+      if SPEC_TEMP_DIR_PID == Process.pid
+        Dir.delete SPEC_TEMP_DIR if File.directory? SPEC_TEMP_DIR
+      end
+    rescue SystemCallError
+      STDERR.puts <<-EOM
+
+  -----------------------------------------------------
+  The rubyspec temp directory is not empty. Ensure that
+  all specs are cleaning up temporary files:
+    #{SPEC_TEMP_DIR}
+  -----------------------------------------------------
+
+      EOM
+    rescue Object => e
+      STDERR.puts "failed to remove spec temp directory"
+      STDERR.puts e.message
     end
-  rescue SystemCallError
-    STDERR.puts <<-EOM
-
------------------------------------------------------
-The rubyspec temp directory is not empty. Ensure that
-all specs are cleaning up temporary files:
-  #{SPEC_TEMP_DIR}
------------------------------------------------------
-
-    EOM
-  rescue Object => e
-    STDERR.puts "failed to remove spec temp directory"
-    STDERR.puts e.message
   end
+else
+  STDERR.puts <<-EOM
+
+-----------------------------------------------------
+                     WARNING
+This Ruby does not support `at_exit`. The spec runner
+will not be able to clean up temporary files.
+-----------------------------------------------------
+
+  EOM
 end
 
 class Object
